Internalise l'état du volet latéral dans le provider plutôt que la page
Next.js 16 impose que les props des composants exportés depuis un fichier 'use client' soient sérialisables (données JSON), sauf les Server Actions (nom action ou suffixe Action). Une fonction client comme onPanelChange déclenche cette règle même entre deux Client Components. En internalisant l’état du panneau (query params) dans le provider, on évite de passer une callback en prop tout en gardant le même comportement (URL synchronisée via ?panel= et ?actionId=).
